+++ tests/test_sklearn_one_hot_encoder_converter.py
@@ -4,17 +4,31 @@
 import unittest
 import packaging.version as pv
 import numpy
-from onnxruntime import __version__ as ort_version
+from numpy.testing import assert_almost_equal
+import pandas
+from onnx.reference import ReferenceEvaluator
+from onnxruntime import InferenceSession, __version__ as ort_version
 from sklearn import __version__ as sklearn_version
 from sklearn.preprocessing import OneHotEncoder
-from sklearn.pipeline import make_pipeline
+from sklearn.pipeline import make_pipeline, Pipeline
 from sklearn.preprocessing import RobustScaler
+from sklearn.compose import ColumnTransformer
+from sklearn.impute import SimpleImputer
+from sklearn.linear_model import LinearRegression
 from skl2onnx import convert_sklearn
 from skl2onnx.common.data_types import (
     Int32TensorType,
     Int64TensorType,
     StringTensorType,
+    FloatTensorType,
 )
+
+try:
+    # scikit-learn >= 0.22
+    from sklearn.utils._testing import ignore_warnings
+except ImportError:
+    # scikit-learn < 0.22
+    from sklearn.utils.testing import ignore_warnings
 from test_utils import dump_data_and_model, TARGET_OPSET
 
 
@@ -88,6 +102,7 @@ class TestSklearnOneHotEncoderConverter(unittest.TestCase):
         not one_hot_encoder_supports_string(),
         reason="OneHotEncoder did not have categories_ before 0.20",
     )
+    @ignore_warnings(category=FutureWarning)
     def test_model_one_hot_encoder_int32_scaler(self):
         model = make_pipeline(
             OneHotEncoder(categories="auto", sparse=False), RobustScaler()
@@ -215,6 +230,7 @@ class TestSklearnOneHotEncoderConverter(unittest.TestCase):
         not one_hot_encoder_supports_string(),
         reason="OneHotEncoder does not support this in 0.19",
     )
+    @ignore_warnings(category=FutureWarning)
     def test_model_one_hot_encoder_list_sparse(self):
         model = OneHotEncoder(
             categories=[[0, 1, 4, 5], [1, 2, 3, 5], [0, 3, 4, 6]], sparse=True
@@ -241,6 +257,7 @@ class TestSklearnOneHotEncoderConverter(unittest.TestCase):
         not one_hot_encoder_supports_string(),
         reason="OneHotEncoder does not support this in 0.19",
     )
+    @ignore_warnings(category=FutureWarning)
     def test_model_one_hot_encoder_list_dense(self):
         model = OneHotEncoder(
             categories=[[0, 1, 4, 5], [1, 2, 3, 5], [0, 3, 4, 6]], sparse=False
@@ -342,6 +359,100 @@ class TestSklearnOneHotEncoderConverter(unittest.TestCase):
             data, model, model_onnx, basename="SklearnOneHotEncoderStringDropFirst2"
         )
 
+    @ignore_warnings(category=RuntimeWarning)
+    def test_shape_inference(self):
+        cat_columns_openings = ["cat_1", "cat_2"]
+        num_columns_openings = [
+            "num_1",
+            "num_2",
+            "num_3",
+            "num_4",
+        ]
+
+        regression_aperturas = LinearRegression()
+
+        numeric_transformer = SimpleImputer(strategy="median")
+        categorical_transformer = Pipeline(
+            steps=[
+                ("imputer", SimpleImputer(strategy="constant", fill_value=-1)),
+                ("onehot", OneHotEncoder(handle_unknown="ignore")),
+            ]
+        )
+
+        preprocessor = ColumnTransformer(
+            transformers=[
+                ("num", numeric_transformer, num_columns_openings),
+                ("cat", categorical_transformer, cat_columns_openings),
+            ]
+        )
+
+        model = Pipeline(
+            steps=[("preprocessor", preprocessor), ("regressor", regression_aperturas)]
+        )
+
+        # Create sample df
+        num_data = numpy.random.rand(100, 4)
+        cat_data = numpy.random.randint(11, size=(100, 2))
+        df = pandas.DataFrame(
+            numpy.hstack((num_data, cat_data)),
+            columns=["num_1", "num_2", "num_3", "num_4", "cat_1", "cat_2"],
+        )
+        df[num_columns_openings] = df[num_columns_openings].astype(float)
+        df[cat_columns_openings] = df[cat_columns_openings].astype(int)
+        df["target"] = numpy.random.rand(100)
+        df["target"] = df["target"].astype(float)
+        X = df.drop("target", axis=1)
+        y = df["target"]
+        model.fit(X, y)
+        X = X[:10]
+        expected = model.predict(X).reshape((-1, 1))
+
+        initial_type = [
+            ("num_1", FloatTensorType([None, 1])),
+            ("num_2", FloatTensorType([None, 1])),
+            ("num_3", FloatTensorType([None, 1])),
+            ("num_4", FloatTensorType([None, 1])),
+            ("cat_1", Int64TensorType([None, 1])),
+            ("cat_2", Int64TensorType([None, 1])),
+        ]
+
+        model_onnx = convert_sklearn(model, initial_types=initial_type)
+
+        feeds = dict(
+            [
+                ("num_1", X.iloc[:, 0:1].values.astype(numpy.float32)),
+                ("num_2", X.iloc[:, 1:2].values.astype(numpy.float32)),
+                ("num_3", X.iloc[:, 2:3].values.astype(numpy.float32)),
+                ("num_4", X.iloc[:, 3:4].values.astype(numpy.float32)),
+                ("cat_1", X.iloc[:, 4:5].values.astype(numpy.int64)),
+                ("cat_2", X.iloc[:, 5:6].values.astype(numpy.int64)),
+            ]
+        )
+
+        # ReferenceEvaluator
+        with self.subTest(engine="onnx"):
+            ref = ReferenceEvaluator(model_onnx)
+            res = ref.run(None, feeds)
+            self.assertEqual(1, len(res))
+            self.assertEqual(expected.shape, res[0].shape)
+            assert_almost_equal(expected, res[0])
+
+        # onnxruntime
+        with self.subTest(engine="onnxruntime"):
+            sess = InferenceSession(
+                model_onnx.SerializeToString(), providers=["CPUExecutionProvider"]
+            )
+            res = sess.run(None, feeds)
+            self.assertEqual(1, len(res))
+            self.assertEqual(expected.shape, res[0].shape)
+            assert_almost_equal(expected, res[0])
+
 
 if __name__ == "__main__":
-    unittest.main()
+    import logging
+
+    for name in ["skl2onnx", "onnx-extended"]:
+        log = logging.getLogger(name)
+        log.setLevel(logging.ERROR)
+    TestSklearnOneHotEncoderConverter().test_model_one_hot_encoder()
+    unittest.main(verbosity=2)
